refactor: passage à l'architecture DDD (Domain-Driven Design)
