+++ src/.env.sample
@@ -1,4 +1,4 @@
-PORT=3003
+PORT_HTTP=3003
 
 PGHOST=localhost
 PGDATABASE=ogalerie
