--- src/api/controllers/index.js
@@ -1,15 +1,0 @@
-'use strict';
-
-import usersController from './users.js';
-import artworksController from './artworks.js';
-import collectionsController from './collections.js';
-import tagsController from './tags.js';
-import commentsController from './comments.js';
-
-export {
-  usersController,
-  collectionsController,
-  artworksController,
-  tagsController,
-  commentsController,
-};
