--- src/api/routers/documentation.js
@@ -1,60 +1,0 @@
-'use strict';
-
-import Router from 'express';
-const router = Router();
-
-import swaggerJSDoc from 'swagger-jsdoc';
-import swaggerUi from 'swagger-ui-express';
-
-const swaggerDefinition = {
-  openapi: '3.1.0',
-  info: {
-    title: 'Express API for Ogalerie',
-    version: '1.0.0',
-    description: 'This is a REST API application made with Express.',
-    license: {
-      name: 'Licensed under MIT',
-      url: 'https://mit-license.org/',
-    },
-    contact: {
-      name: 'Ogalerie',
-      url: 'https://github.com/O-clock-Yost/projet-06-o-galerie-back',
-    },
-  },
-  servers: [
-    {
-      url: 'http://localhost:3003',
-      description: 'Development server',
-    },
-    {
-      url: 'http://arnaudpithon-server.eddi.cloud:8080',
-      description: 'Production Server OClock',
-    },
-  ],
-  components: {
-    securitySchemes: {
-      bearerAuth: {
-        type: 'http',
-        scheme: 'bearer',
-        bearerFormat: 'JWT',
-      },
-    },
-  },
-  security: [
-    {
-      bearerAuth: [],
-    },
-  ],
-};
-
-const options = {
-  swaggerDefinition,
-  // Paths to files containing OpenAPI definitions
-  apis: ['./api/routers/*.js'],
-};
-
-const swaggerSpec = swaggerJSDoc(options);
-
-router.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
-
-export default router;
