+++ src/app.js
@@ -1,0 +1,3 @@
+import { startServer } from './interfaces/http/server.js';
+
+startServer();
