+++ src/config/db.js
@@ -1,0 +1,5 @@
+import 'dotenv/config';
+
+const { PGUSER, PGPASSWORD, PGHOST, PGDATABASE } = process.env;
+
+export default { PGUSER, PGPASSWORD, PGHOST, PGDATABASE };
