+++ src/modules/comments/router.js
@@ -1,0 +1,39 @@
+import Router from 'express';
+
+const router = Router();
+
+import securityService from '../../middlewares/security.js';
+
+import commentsController from './controller.js';
+
+/**
+ * @swagger
+ * /v1/comments/{id}:
+ *   patch:
+ *      summary: Modify a comment
+ *      tags:
+ *          - comments
+ */
+router.patch(
+  '/:id',
+  securityService.isConnected,
+  securityService.isCommentOwner,
+  commentsController.update,
+);
+
+/**
+ * @swagger
+ * /v1/comments/{id}:
+ *   delete:
+ *      summary: Delete a comment
+ *      tags:
+ *          - comments
+ */
+router.delete(
+  '/:id',
+  securityService.isConnected,
+  securityService.isCommentOwner,
+  commentsController.delete,
+);
+
+export default router;
