+++ src/vite.config.js
@@ -1,5 +1,6 @@
 export default {
   test: {
     environment: 'node',
+    include: ['**/*.test.js', '__tests__/**/*.js'],
   },
 };
